// src/client/services/service-proxies.ts
export interface User {
    _id: string;
    email: string;
    familyName: string;
    givenName: string;
    fullName: string;
    phoneNo: string;
    address: string;
    description: string;
    avatarUrl: string;
    province: string;
    dob: number;
    gender: string;
    loginDetail: any;
    roles: string[];
    completeSignUp: boolean;
    followBy: any[];
    isActive: boolean;
    createdAt: number;
    createdBy: number;
    counterFollow: number;
    isFollowing: boolean;
}
// ...
export interface CommonKeyword {
    keyword: string;
    count: number;
    createdAt: number;
}

export interface GetCommonKeywordsResult {
    data: CommonKeyword[];
    before: string;
    after: string;
}

export interface CreateRecentViewedPayload {
    news: string;
}

export interface RecentViewedItem {
    _id: string;
    uuid: string;
    user: string;
    news: any;
    createdAt: number;
    lastModifiedAt: number;
}
// ...
export interface ElasticSearchResult {
    data: ElasticSearchItemResult[];
    before: string;
    after: string;
}

export interface ElasticSearchOffsetResult {
    data: ElasticSearchItemResult[];
    total: number;
}

export interface ElasticSearchItemResult {
    title: string;
    state: string;
    description: string;
    shop: any;
    imageUrls: string[];
    price: number;
    priceType: string;
    newsType: string;
    owner: any;
    location: Location;
    status: string;
    usedHours: number;
    serialNo: number;
    categoryId: CategoryId;
    brand: Brand2;
    model: Model2;
    weight: number;
    year: number;
    followBy: string[];
    isDeleted: boolean;
    createdBy: number;
    createdAt: number;
    lastModifiedAt: number;
    elasticsearchID: string;
    elasticsearchTitle: string;
    isFollowing: boolean;
    _id: string;
}
// ...
export interface Brand {
    _id: string;
    slug: string;
    name: string;
    logoImage: string;
    totalNews: number;
    brandCategories: Category[];
    brandModels: Model[];
    createdAt: number;
}

export interface FindBrandsResult {
    data: Brand[];
    before: string;
    after: string;
}

export interface CreateBrandPayload {
    name: string;
    logoImage: string;
}

export interface UpdateBrandPayload {
    operation: UpdateBrandPayloadOperation;
    payload: any;
}

export interface Category {
    _id: string;
    slug: string;
    name: string;
    imageUrl: string;
    totalNews: number;
    categoryBrands: Brand[];
    categoryModels: Model[];
    createdAt: number;
}

export interface FindCategoriesResult {
    data: Category[];
}

export interface CreateCategoryPayload {
    name: string;
    imageUrl: string;
}

export interface UpdateCategoryPayload {
    operation: UpdateCategoryPayloadOperation;
    payload: Payload13;
}

export interface Model {
    _id: string;
    slug: string;
    name: string;
    brand: any;
    category: any;
    totalNews: number;
    createdAt: number;
}

export interface FindModelsResult {
    data: Model[];
    before: string;
    after: string;
}

export interface CreateModelPayload {
    name: string;
    slug: string;
    brand: string;
    model: string;
}

export interface UpdateModelPayload {
    _id: string;
    name: string;
    slug: string;
    brand: string;
    model: string;
}

export interface Province {
    _id: string;
    slug: string;
    name: string;
    totalNews: number;
}
// ...
export interface Location {
    isDeleted: boolean;
    lastModifiedAt: number;
    _id: string;
    name: string;
    slug: string;
}

export interface CategoryId {
    isDeleted: boolean;
    lastModifiedAt: number;
    _id: string;
    name: string;
    slug: string;
}

export interface Brand2 {
    isDeleted: boolean;
    lastModifiedAt: number;
    _id: string;
    name: string;
    slug: string;
    totalNews: number;
}

export interface Model2 {
    isDeleted: boolean;
    lastModifiedAt: number;
    _id: string;
    name: string;
    slug: string;
    brand: string;
    category: string;
}